+++ src/components/sections/InventorySection.component.jsx
@@ -1,23 +1,20 @@
 import ItemCard from "../cards/ItemCard.component";
 import EmptyState from "../ui/EmptyState.component";
 
-function InventorySection({id, title, items}) {
+function InventorySection({ id, title, items, shoppingCart = false }) {
   return (
-      <section id={id}>
-        <h2>{title}</h2>
-        <ul>
-          {items && items.length > 0 ? (
-            items.map((item) => (
-              <ItemCard
-                key={item.id}
-                item={item}
-              />
-            ))
-          ) : (
-            <EmptyState title={title.toLowerCase()} />
-          )}
-        </ul>
-      </section>
+    <section id={id}>
+      <h2>{title}</h2>
+      <ul>
+        {items && items.length > 0 ? (
+          items.map((item) => (
+            <ItemCard key={item.id} item={item} shoppingCart={shoppingCart} />
+          ))
+        ) : (
+          <EmptyState title={title.toLowerCase()} />
+        )}
+      </ul>
+    </section>
   );
 }